Modified subscription_plan validation.
Previously it attempted to validate the presence of a
"redemption_key",
but the SubscriptionPlan object does not have or need a "redemption_key"
attribute. The "redemption_key" attribute belongs to the Coupon model.
SubscriptionPlan does have a "key" attribute, so I'm validating this
instead. (Though technically there's no reason it's required.)

	modified:   lib/freemium/subscription_plan.rb