Move billing_key from the Subscription model to the CreditCard mode.
The billing_key is the token that we use to charge a credit card
stored at our payment processor. It used to be stored as an attribute
of the Subscription model, which didn't make a lot of sense to
me. This update moves it to the billing_key to the CreditCard model.

This change will make it much easier to update credit card
information. One can simply use a standard form_for credit_card with
standard validations.

This commit updates the create_freemium_models generator to use the
new schema. If you have an existing database that uses the old schema
you can use the migration in
script/move_billing_key_from_subscription_to_credit_card.rb .

	modified:   generators/freemium_migration/templates/migration.rb
	modified:   init.rb
	modified:   lib/freemium/credit_card.rb
	modified:   lib/freemium/subscription.rb
	new file:   script/move_billing_key_from_subscription_to_credit_card.rb