Numerous changes to make me less afraid of running a billing cycle.
Improved documentation in places.

Fixed assignment problems that could result with Transaction and
CouponRedemption objects when 'attr_accessible' is enabled (meaning,
confusingly, that attributes are _not_ accessible by default).

Added a configurable logging object which, by default, uses the Rails
logger.

Log all HTTP post and responses to the payment processor.

Save the 'message' in the Braintree response into the Transaction object.

Added Hoptoad notifications to all caught exceptions.

Renamed 'invoice' in the SubscriptionMailer to 'payment_receipt', which
more accurately describes its contents.

Deleted the set_paid_through method. Honestly, I’m not totally sure
what it was supposed to do. It was a maze of conditionals. The
important part of the function, which is to set the in_trial flag and
paid_through date for new accounts, has been moved to a callback:
before_validation_on_create :start_free_trial