Fixed a few CreditCard issues.
When the 'billing_key' was moved from the Subscription model to the
CreditCard model, the function function to delete the credit card
entry from the payment processor ('cancel_in_remote_system') was
accidently left with the Subscription model. This has been
corrected. The cancel_in_remote_system method will be called whenever

1) A subscription changes from an paid to an unpaid plan
2) The credit card number is updated (immediately before the new
number is stored)
3) The credit card object is destroyed

This commit also replaces the implementation of expiration_date
attribute and accessors, which were super manky and failed on a number
of mundane cases.