Fixed an issue that caused CreditCard to look invalid.
The 'changed?' method in credit_card.rb tests each of the virtual (ie,
non-database) attributes of the CreditCard object, and returns 'true'
if any of the virtual attributes exist. The 'month' and 'year'
accessors were modified to return the values from 'expiration_date',
and this change made it look like any CreditCard object with an
expiration_date was always changed. As a result, valid already-saved
CreditCards looked invalid.

This commit also improves logging and notification in the event that
the Subscription object does not update correctly after a transaction
is processed.